chore: break files into smaller componets
